refactor(specialist): split editor into form hook and capabilities section
- useSpecialistEditorForm owns the form state machine: mount-time
  seeding (profile, create prefill, restorable draft) and the
  unsaved-form draft kept in the specialist store.
  formFromProfile/formFromCreateInput replace the hand-repeated seed
  literals (Reload now reuses the profile seed), and the form payload
  type aliases the store's SpecialistEditorFormDraft instead of a
  parallel copy
- SpecialistCapabilitiesSection owns the capabilities block end to end:
  full-access switch, tabs, add dropdowns, selected rows with inline
  navigation, and the catalog row memos; fully controlled via
  capabilityMode / id arrays / updater callbacks so the editor's form
  remains the single source of truth

Pure move: no behavior, UI, or test changes.
